+++ tests/test_execute.py
@@ -1,9 +1,11 @@
 import shutil
+from unittest import mock
 
 import pytest
 
 from jupytext import read
 from jupytext.cli import jupytext
+from jupytext.version import __version__
 
 from .utils import requires_ir_kernel, requires_nbconvert, skip_on_windows
 
@@ -227,3 +229,48 @@ sum(ast.literal_eval(line) for line in text.splitlines())
     nb = read(tmp2_ipynb)
     assert len(nb.cells) == 3
     assert nb.cells[2].outputs[0]["data"] == {"text/plain": "3"}
+
+
+@pytest.fixture()
+def sample_md_notebook():
+    """This is a sample md notebook with an outdated version of Jupytext
+    and no kernel information, to test #908"""
+    return """---
+jupyter:
+  jupytext:
+    text_representation:
+      extension: .md
+      format_name: markdown
+      format_version: '1.1'
+      jupytext_version: 1.1.0
+---
+
+```python
+1 + 1
+```
+"""
+
+
+def test_execute_text_file_does_update_the_metadata(sample_md_notebook, tmp_path):
+    md_file = tmp_path / "nb.md"
+    md_file.write_text(sample_md_notebook)
+
+    jupytext([str(md_file), "--execute"])
+
+    new_md_text = md_file.read_text()
+    assert __version__ in new_md_text
+    assert "kernelspec" in new_md_text
+
+
+def test_cat_execute_does_not_update_the_metadata(sample_md_notebook, tmp_path):
+    md_file = tmp_path / "nb.md"
+    md_file.write_text(sample_md_notebook)
+
+    # read md notebook on stdin - this does the same as
+    # cat notebook.md | jupytext --execute
+    with open(md_file) as fp, mock.patch("sys.stdin", fp):
+        jupytext(["--execute"])
+
+    new_md_text = md_file.read_text()
+    assert __version__ not in new_md_text
+    assert "kernelspec" not in new_md_text